backend/requirements.txt
# Основные зависимости
fastapi==0.110.0
uvicorn[standard]==0.29.0
qdrant-client==1.9.0
httpx>=0.25.2,<0.28.0  # Разрешаем диапазон версий
python-dotenv==1.0.1
pydantic==2.6.0
python-multipart==0.0.9

# Явно устанавливаем CPU-версию PyTorch ПЕРВОЙ
torch==2.9.0 --index-url https://download.pytorch.org/whl/cpu

# Остальные зависимости
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.13.0
huggingface-hub==0.20.3
tokenizers==0.15.2
transformers==4.38.0

# Для документов
python-docx==1.1.0
PyPDF2==3.0.1
pdf2image==1.17.0

# Интеграции (Telegram + Bitrix24)
python-telegram-bot==20.7
aiohttp==3.9.0